backend/requirements.txt

# Core dependencies
pydantic>=2.10.0
pydantic-ai>=0.2.4
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0

# Supabase integration
supabase>=1.0.3  # Python SDK for Supabase

# API client for backend interactions
httpx>=0.24.1    # Async HTTP client used by FastAPI

# CORS middleware
fastapi-cors>=0.0.6  # For cross-origin requests between frontend and backend

# OpenAI integration
openai>=1.3.0    # OpenAI API client for GPT-4o integration

# Utility packages
python-multipart>=0.0.6  # For handling form data
email-validator>=2.0.0   # For email validation

# Database connectivity
# Note: Supabase handles PostgreSQL connections, but we include this for direct DB operations if needed
asyncpg>=0.29.0  # Async PostgreSQL driver

# Logging
loguru>=0.7.0    # Better logging

# Production dependencies
gunicorn>=21.2.0  # WSGI HTTP Server for production deployment
uvloop>=0.18.0    # Fast event loop for uvicorn (Linux/MacOS only)
httptools>=0.6.0  # Fast HTTP parser

# Testing dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.24.1